add entity, service, controllers, flyway fixed mistakes

+++ src/main/java/com/dataox/dto/JobPostingDTO.java
@@ -3,7 +3,6 @@ package com.dataox.dto;
 import lombok.Getter;
 import lombok.Setter;
 import lombok.ToString;
-
 import java.util.List;
 
 @Getter
@@ -17,8 +16,7 @@ public class JobPostingDTO {
     private String organizationTitle;
     private String laborFunction;
     private List<String> locations;
-    private int postedDateUnix;
+    private long postedDateUnix;
     private String descriptionHtml;
     private List<String> tags;
-
 }
